LongNoteSample.cs 수정
LongNoteSample에 있던 Time 속성을 제거하고, NotePosition 속성에 있는 Time의 값을 사용함에 따른 변경사항 적용.

+++ Assets/Scripts/Beatale/ChartSystem/LongNoteSample.cs
@@ -4,9 +4,9 @@
     {
         public float Degree;
         public float Width;
-        public float Time;
+        public double Time;
 
-        public LongNoteSample(float degree, float width, float time)
+        public LongNoteSample(float degree, float width, double time)
         {
             Degree = degree;
             Width = width;
